v0.8.0 (April 1st, 2020) - Added flanking and menu options to configure (Adv/+1/+2/+5). Only works for medium or smaller targets and attackers. - Added menu options to switch between key combination of "Alt + L-Click" or "Shift + Alt + L-Click", for deleting tokens from map. - Added exit clause for automatic ranged weapon attacks from PC sheets, for non-standardized range entries.

+++ scripts/token_remover.lua
@@ -13,18 +13,40 @@ end
 -- Post: Token removed from combat map (Alt) or combat map and combat tracker (Atl+Ctrl)
 -- NOTE: button (number), Returns a numerical value indicating the button pressed (1 = left, 2 = middle, 4 = button 4, 5 = button 5). Right button is used for radial menus.
 function onClickDown( token, button, image ) 
-	-- Deletes token from combat map, if Alt held on left mouse click.
-	if (Input.isAltPressed() == true) and (User.isHost() == true) and (button==1) then
-		local nodeCT = CombatManager.getCTFromToken(token);
-		token.delete();		
 
-		-- Deletes token from combat tracker if Ctrl was also held on click.
-		if (Input.isControlPressed() == true) then
-			if nodeCT then -- only attempt delete if there is a CT entry
-				nodeCT.delete();					
-			end				
-		end
-	end	
+	local bFlankingRules = OptionsManager.getOption("CE_TRBC"); 
+
+	-- Deletes token from combat map, if Alt held on left mouse click.
+	if (bFlankingRules == 'option_val_alt') then	
+		if (Input.isAltPressed() == true) and (User.isHost() == true) and (button==1) then
+			local nodeCT = CombatManager.getCTFromToken(token);
+			token.delete();		
+	
+			-- Deletes token from combat tracker if Ctrl was also held on click.
+			if (Input.isControlPressed() == true) then
+				if nodeCT then -- only attempt delete if there is a CT entry
+					nodeCT.delete();					
+				end				
+			end
+		end	
+	end
+
+	-- Deletes token from combat map, if Alt + Shift held on left mouse click.
+	if (bFlankingRules == 'option_val_alt_shift') then	
+		if (Input.isAltPressed() == true) and (Input.isShiftPressed() == true) and (User.isHost() == true) and (button==1) then
+			local nodeCT = CombatManager.getCTFromToken(token);
+			token.delete();		
+	
+			-- Deletes token from combat tracker if Ctrl was also held on click.
+			if (Input.isControlPressed() == true) then
+				if nodeCT then -- only attempt delete if there is a CT entry
+					nodeCT.delete();					
+				end				
+			end
+		end	
+	end
+
+
 end
